+++ package.json
@@ -2,14 +2,19 @@
   "name": "ipld",
   "version": "0.4.0",
   "description": "JavaScript implementation of the IPLD (InterpPlanetary Linked Data)",
-  "main": "src/index.js",
+  "main": "lib/index.js",
+  "jsnext:main": "src/index.js",
   "scripts": {
-    "test": "./node_modules/.bin/lab tests/*-test.js",
-    "codestyle-fix": "./node_modules/.bin/standard --format",
-    "codestyle": "./node_modules/.bin/standard"
+    "lint": "dignified-lint",
+    "build": "dignified-build",
+    "test": "dignified-test",
+    "test:node": "dignified-test node",
+    "test:browser": "dignified-test browser",
+    "release": "dignified-release"
   },
   "pre-commit": [
-    "codestyle"
+    "lint",
+    "test"
   ],
   "repository": {
     "type": "git",
@@ -35,9 +40,9 @@
     "underscore": "^1.8.3"
   },
   "devDependencies": {
+    "chai": "^3.5.0",
     "code": "^1.5.0",
-    "lab": "^5.15.2",
-    "pre-commit": "^1.1.1",
-    "standard": "^5.1.0"
+    "dignified.js": "github:dignifiedquire/dignified.js",
+    "pre-commit": "^1.1.1"
   }
 }
